+++ R/density_ratios.R
@@ -1,18 +1,20 @@
-cf_r <- function(task, learners, mtp, control, progress_bar) {
-  out <- list()
+cf_r <- function(task, learners, mtp, control, pb) {
+  out <- vector("list", length = length(task$folds))
   for (fold in seq_along(task$folds)) {
     out[[fold]] <- future::future({
-      estimate_r(get_folded_data(task$natural, task$folds, fold),
-                 get_folded_data(task$shifted, task$folds, fold),
-                 task$trt,
-                 task$cens,
-                 task$risk,
-                 task$tau,
-                 task$node_list$trt,
-                 learners,
-                 mtp,
-                 control,
-                 progress_bar)
+      estimate_r(
+        get_folded_data(task$natural, task$folds, fold),
+        get_folded_data(task$shifted, task$folds, fold),
+        task$trt,
+        task$cens,
+        task$risk,
+        task$tau,
+        task$node_list$trt,
+        learners,
+        pb,
+        mtp,
+        control
+      )
     },
     seed = TRUE)
   }
@@ -20,9 +22,9 @@ cf_r <- function(task, learners, mtp, control, progress_bar) {
   trim_ratios(recombine_ratios(future::value(out), task$folds), control$.trim)
 }
 
-estimate_r <- function(natural, shifted, trt, cens, risk, tau, node_list, learners, mtp, control, progress_bar) {
+estimate_r <- function(natural, shifted, trt, cens, risk, tau, node_list, learners, pb, mtp, control) {
   densratios <- matrix(nrow = nrow(natural$valid), ncol = tau)
-  fits <- list()
+  fits <- vector("list", length = tau)
 
   for (t in 1:tau) {
     jrt <- rep(censored(natural$train, cens, t)$j, 2)
@@ -31,9 +33,13 @@ estimate_r <- function(natural, shifted, trt, cens, risk, tau, node_list, learne
     jrv <- censored(natural$valid, cens, t)$j
     drv <- at_risk(natural$valid, risk, t)
 
-    trt_t <- ifelse(length(trt) > 1, trt[t], trt)
+    if (length(trt) > 1) {
+      trt_t <- trt[[t]]
+    } else {
+      trt_t <- trt[[1]]
+    }
 
-    frv <- followed_rule(natural$valid[[trt_t]], shifted$valid[[trt_t]], mtp)
+    frv <- followed_rule(natural$valid[, trt_t], shifted$valid[, trt_t], mtp)
 
     vars <- c(node_list[[t]], cens[[t]])
     stacked <- stack_data(natural$train, shifted$train, trt, cens, t)
@@ -52,7 +58,7 @@ estimate_r <- function(natural, shifted, trt, cens, risk, tau, node_list, learne
     }
 
     pred <- matrix(-999L, nrow = nrow(natural$valid), ncol = 1)
-    pred[jrv & drv, ] <- bound(SL_predict(fit, natural$valid[jrv & drv, ]), .Machine$double.eps)
+    pred[jrv & drv, ] <- bound(SL_predict(fit, natural$valid[jrv & drv, c("lmtp_id", vars)]), .Machine$double.eps)
 
     ratios <- density_ratios(pred, irv, drv, frv, mtp)
     densratios[, t] <- ratios
@@ -67,7 +73,7 @@ stack_data <- function(natural, shifted, trt, cens, tau) {
   shifted_half <- natural
 
   if (length(trt) > 1 || tau == 1) {
-    shifted_half[[trt[tau]]] <- shifted[[trt[tau]]]
+    shifted_half[, trt[[tau]]] <- shifted[, trt[[tau]]]
   }
 
   if (!is.null(cens)) {