+++ R/estimators.R
@@ -7,8 +7,11 @@
 #' @param data \[\code{data.frame}\]\cr
 #'  A \code{data.frame} in wide format containing all necessary variables
 #'  for the estimation problem. Must not be a \code{data.table}.
-#' @param trt \[\code{character}\]\cr
+#' @param trt \[\code{character}\] or \[\code{list}\]\cr
 #'  A vector containing the column names of treatment variables ordered by time.
+#'  Or, a list of vectors, the same length as the number of time points of observation.
+#'  Vectors should contain column names for the treatment variables at each time point. The list
+#'  should be ordered following the time ordering of the model.
 #' @param outcome \[\code{character}\]\cr
 #'  The column name of the outcome variable. In the case of time-to-event
 #'  analysis, a vector containing the columns names of intermediate outcome variables and the final
@@ -50,17 +53,19 @@
 #'  An optional, ordered vector of the bounds for a continuous outcomes. If \code{NULL},
 #'  the bounds will be taken as the minimum and maximum of the observed data.
 #'  Should be left as \code{NULL} if the outcome type is binary.
-#' @param learners_outcome \[\code{character}\]\cr
-#' @param learners_trt \[\code{character}\]\cr \bold{Only include candidate learners capable of binary classification}.
-#' @param trt_method \[\code{character}\]\cr
-#'  Method for estimating treatment assignment mechanism (default or riesz)
+#' @param learners_outcome \[\code{character}\]\cr A vector of \code{mlr3superlearner} algorithms for estimation
+#'  of the outcome regression. Default is \code{c("mean", "glm")}.
+#' @param learners_trt \[\code{character}\]\cr A vector of \code{mlr3superlearner} algorithms for estimation
+#'  of the outcome regression. Default is \code{c("mean", "glm")}.
+#'  \bold{Only include candidate learners capable of binary classification}.
+#' @param riesz \[\code{logical(1)}\]\cr
+#'  Estimate treatment mechanism using riesz representer? Default is \code{FALSE}.
 #' @param folds \[\code{integer(1)}\]\cr
 #'  The number of folds to be used for cross-fitting.
 #' @param weights \[\code{numeric(nrow(data))}\]\cr
 #'  An optional vector containing sampling weights.
 #' @param control \[\code{list()}\]\cr
 #'  Output of \code{lmtp_control()}.
-#' @param ... Extra arguments. Exists for backwards compatibility.
 #'
 #' @details
 #' ## Should \code{mtp = TRUE}?
@@ -97,9 +102,8 @@ lmtp_tmle <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
                       id = NULL, bounds = NULL,
                       learners_outcome = c("mean", "glm"),
                       learners_trt = c("mean", "glm"),
-                      trt_method = "default",
-                      folds = 10, weights = NULL,
-                      control = lmtp_control(), ...) {
+                      riesz = FALSE, folds = 10, weights = NULL,
+                      control = lmtp_control()) {
 
   assertNotDataTable(data)
   checkmate::assertCharacter(outcome, len = if (match.arg(outcome_type) != "survival") 1,
@@ -108,11 +112,11 @@ lmtp_tmle <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
 
   tau <- determine_tau(outcome, trt)
 
-  assertTrtCharacter(trt, tau)
+  assert_trt(trt, tau)
   checkmate::assertCharacter(cens, len = tau, null.ok = !checkmate::anyMissing(data[, outcome, drop = FALSE]))
   checkmate::assertList(time_vary, types = c("NULL", "character"), len = tau, null.ok = TRUE)
   checkmate::assertCharacter(id, len = 1, null.ok = TRUE)
-  checkmate::assertSubset(c(trt, outcome, baseline, unlist(time_vary), cens, id), names(data))
+  checkmate::assertSubset(c(unlist(trt), outcome, baseline, unlist(time_vary), cens, id), names(data))
   assertLmtpData(data, trt, outcome, baseline, time_vary, cens, id)
   assertOutcomeTypes(data, outcome, match.arg(outcome_type))
   assertReservedNames(data)
@@ -122,69 +126,69 @@ lmtp_tmle <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
   checkmate::assertNumeric(weights, len = nrow(data), finite = TRUE, any.missing = FALSE, null.ok = TRUE)
   checkmate::assertNumber(k, lower = 0, upper = Inf)
   checkmate::assertNumber(folds, lower = 1, upper = nrow(data) - 1)
-  checkmate::assertSubset(c(trt, outcome, baseline, unlist(time_vary), cens, id), names(data))
   checkmate::assertNumber(control$.learners_outcome_folds, null.ok = TRUE)
   checkmate::assertNumber(control$.learners_trt_folds, null.ok = TRUE)
+  checkmate::assertNumber(control$.bound)
   checkmate::assertNumber(control$.trim, upper = 1)
   checkmate::assertLogical(control$.return_full_fits, len = 1)
+  check_trt_type(data, unlist(trt), mtp)
   checkmate::assertMatrix(conditional, mode = c("logical"), null.ok = TRUE)
 
-  if(!is.null(conditional) && trt_method == "default") {
-    stop("For conditional estimation, trt_method must be set to `riesz`.")
+  if (!is.null(conditional) && isFALSE(riesz) {
+    stop("For conditional estimation, set `riesz = TRUE`.")
   }
 
-  extras <- list(...)
-  if ("intervention_type" %in% names(extras)) {
-    mtp <- extras$intervention_type == "mtp"
-    warning("The `intervention_type` argument of `lmtp_tmle()` is deprecated as of lmtp 1.3.1",
-            call. = FALSE)
+  task <- lmtp_task$new(
+    data = data,
+    trt = trt,
+    outcome = outcome,
+    time_vary = time_vary,
+    baseline = baseline,
+    cens = cens,
+    k = k,
+    shift = shift,
+    shifted = shifted,
+    id = id,
+    outcome_type = match.arg(outcome_type),
+    V = folds,
+    weights = weights,
+    bounds = bounds,
+    bound = control$.bound
+  )
+
+  pb <- progressr::progressor(task$tau*folds*2)
+      
+  if (isFALSE(riesz) {
+    ratios <- cf_r(task, learners_trt, mtp, control, pb)
+  } else {
+    ratios <- cf_rr(task, learners_trt, mtp, control, pb)
   }
 
-  Task <- lmtp_Task$new(data = data,
-                        trt = trt,
-                        outcome = outcome,
-                        time_vary = time_vary,
-                        baseline = baseline,
-                        cens = cens,
-                        conditional = conditional,
-                        k = k,
-                        shift = shift,
-                        shifted = shifted,
-                        id = id,
-                        outcome_type = match.arg(outcome_type),
-                        V = folds,
-                        weights = weights,
-                        bounds = bounds)
-
-  progress_bar <- progressr::progressor(Task$tau*folds*2)
-
-  if(trt_method == "default") {
-    ratios <- cf_r(Task, learners_trt, mtp, control, progress_bar)
-  }
-  else {
-    ratios <- cf_rr(Task, learners_trt, mtp, control, progress_bar)
-  }
-
-  estims <- cf_tmle(Task, "tmp_lmtp_scaled_outcome", trt_method == "riesz", ratios$ratios,
-                    learners_outcome, control, progress_bar)
+  estims <- cf_tmle(task,
+                    "tmp_lmtp_scaled_outcome",
+                    riesz,
+                    ratios$ratios,
+                    learners_outcome,
+                    control,
+                    pb)
 
   theta_dr(
     list(
       estimator = "TMLE",
       m = list(natural = estims$natural, shifted = estims$shifted),
       r = ratios$ratios,
-      cumulated = trt_method == "riesz",
-      tau = Task$tau,
-      folds = Task$folds,
-      id = Task$natural$lmtp_id,
-      outcome_type = Task$outcome_type,
-      bounds = Task$bounds,
-      weights = Task$weights,
+      cumulated = riesz,
+      tau = task$tau,
+      folds = task$folds,
+      id = task$id,
+      outcome_type = task$outcome_type,
+      bounds = task$bounds,
+      weights = task$weights,
       shift = if (is.null(shifted)) deparse(substitute((shift))) else NULL,
       fits_m = estims$fits,
       fits_r = ratios$fits,
-      outcome_type = Task$outcome_type,
-      conditional = Task$conditional
+      outcome_type = task$outcome_type, 
+      conditional = task$conditional
     ),
     FALSE
   )
@@ -199,8 +203,11 @@ lmtp_tmle <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
 #' @param data \[\code{data.frame}\]\cr
 #'  A \code{data.frame} in wide format containing all necessary variables
 #'  for the estimation problem. Must not be a \code{data.table}.
-#' @param trt \[\code{character}\]\cr
+#' @param trt \[\code{character}\] or \[\code{list}\]\cr
 #'  A vector containing the column names of treatment variables ordered by time.
+#'  Or, a list of vectors, the same length as the number of time points of observation.
+#'  Vectors should contain column names for the treatment variables at each time point. The list
+#'  should be ordered following the time ordering of the model.
 #' @param outcome \[\code{character}\]\cr
 #'  The column name of the outcome variable. In the case of time-to-event
 #'  analysis, a vector containing the columns names of intermediate outcome variables and the final
@@ -238,15 +245,17 @@ lmtp_tmle <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
 #'  An optional, ordered vector of the bounds for a continuous outcomes. If \code{NULL},
 #'  the bounds will be taken as the minimum and maximum of the observed data.
 #'  Should be left as \code{NULL} if the outcome type is binary.
-#' @param learners_outcome \[\code{character}\]\cr
-#' @param learners_trt \[\code{character}\]\cr \bold{Only include candidate learners capable of binary classification}.
+#' @param learners_outcome \[\code{character}\]\cr A vector of \code{mlr3superlearner} algorithms for estimation
+#'  of the outcome regression. Default is \code{c("mean", "glm")}.
+#' @param learners_trt \[\code{character}\]\cr A vector of \code{mlr3superlearner} algorithms for estimation
+#'  of the outcome regression. Default is \code{c("mean", "glm")}.
+#'  \bold{Only include candidate learners capable of binary classification}.
 #' @param folds \[\code{integer(1)}\]\cr
 #'  The number of folds to be used for cross-fitting.
 #' @param weights \[\code{numeric(nrow(data))}\]\cr
 #'  An optional vector containing sampling weights.
 #' @param control \[\code{list()}\]\cr
 #'  Output of \code{lmtp_control()}.
-#' @param ... Extra arguments. Exists for backwards compatibility.
 #'
 #' @details
 #' ## Should \code{mtp = TRUE}?
@@ -285,7 +294,7 @@ lmtp_sdr <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
                      learners_outcome = c("mean", "glm"),
                      learners_trt = c("mean", "glm"),
                      folds = 10, weights = NULL,
-                     control = lmtp_control(), ...) {
+                     control = lmtp_control()) {
 
   assertNotDataTable(data)
   checkmate::assertCharacter(outcome, len = if (match.arg(outcome_type) != "survival") 1,
@@ -294,11 +303,11 @@ lmtp_sdr <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
 
   tau <- determine_tau(outcome, trt)
 
-  assertTrtCharacter(trt, tau)
+  assert_trt(trt, tau)
   checkmate::assertCharacter(cens, len = tau, null.ok = !checkmate::anyMissing(data[, outcome, drop = FALSE]))
   checkmate::assertList(time_vary, types = c("NULL", "character"), len = tau, null.ok = TRUE)
   checkmate::assertCharacter(id, len = 1, null.ok = TRUE)
-  checkmate::assertSubset(c(trt, outcome, baseline, unlist(time_vary), cens, id), names(data))
+  checkmate::assertSubset(c(unlist(trt), outcome, baseline, unlist(time_vary), cens, id), names(data))
   assertLmtpData(data, trt, outcome, baseline, time_vary, cens, id)
   assertOutcomeTypes(data, outcome, match.arg(outcome_type))
   assertReservedNames(data)
@@ -308,39 +317,39 @@ lmtp_sdr <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
   checkmate::assertNumeric(weights, len = nrow(data), finite = TRUE, any.missing = FALSE, null.ok = TRUE)
   checkmate::assertNumber(k, lower = 0, upper = Inf)
   checkmate::assertNumber(folds, lower = 1, upper = nrow(data) - 1)
-  checkmate::assertSubset(c(trt, outcome, baseline, unlist(time_vary), cens, id), names(data))
   checkmate::assertNumber(control$.learners_outcome_folds, null.ok = TRUE)
   checkmate::assertNumber(control$.learners_trt_folds, null.ok = TRUE)
+  checkmate::assertNumber(control$.bound)
   checkmate::assertNumber(control$.trim, upper = 1)
   checkmate::assertLogical(control$.return_full_fits, len = 1)
+  check_trt_type(data, unlist(trt), mtp)
 
-  Task <- lmtp_Task$new(data = data,
-                        trt = trt,
-                        outcome = outcome,
-                        time_vary = time_vary,
-                        baseline = baseline,
-                        cens = cens,
-                        k = k,
-                        shift = shift,
-                        shifted = shifted,
-                        id = id,
-                        outcome_type = match.arg(outcome_type),
-                        V = folds,
-                        weights = weights,
-                        bounds = bounds)
+  task <- lmtp_task$new(
+    data = data,
+    trt = trt,
+    outcome = outcome,
+    time_vary = time_vary,
+    baseline = baseline,
+    cens = cens,
+    k = k,
+    shift = shift,
+    shifted = shifted,
+    id = id,
+    outcome_type = match.arg(outcome_type),
+    V = folds,
+    weights = weights,
+    bounds = bounds,
+    bound = control$control$.bound
+  )
 
-  extras <- list(...)
-  if ("intervention_type" %in% names(extras)) {
-    mtp <- extras$intervention_type == "mtp"
-    warning("The `intervention_type` argument of `lmtp_sdr()` is deprecated as of lmtp 1.3.1",
-            call. = FALSE)
-  }
+  pb <- progressr::progressor(task$tau*folds*2)
 
-  progress_bar <- progressr::progressor(Task$tau*folds*2)
-
-  ratios <- cf_r(Task, learners_trt, mtp, control, progress_bar)
-  estims <- cf_sdr(Task, "tmp_lmtp_scaled_outcome", ratios$ratios,
-                   learners_outcome, control, progress_bar)
+  ratios <- cf_r(task, learners_trt, mtp, control, pb)
+  estims <- cf_sdr(task,
+                   "tmp_lmtp_scaled_outcome",
+                   ratios$ratios,
+                   learners_outcome,
+                   control,
 
   theta_dr(
     list(
@@ -348,17 +357,17 @@ lmtp_sdr <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
       m = list(natural = estims$natural, shifted = estims$shifted),
       r = ratios$ratios,
       cumulated = FALSE,
-      tau = Task$tau,
-      folds = Task$folds,
-      id = Task$natural$lmtp_id,
-      outcome_type = Task$outcome_type,
-      bounds = Task$bounds,
-      weights = Task$weights,
+      tau = task$tau,
+      folds = task$folds,
+      id = task$id,
+      outcome_type = task$outcome_type,
+      bounds = task$bounds,
+      weights = task$weights,
       shift = if (is.null(shifted)) deparse(substitute((shift))) else NULL,
       fits_m = estims$fits,
       fits_r = ratios$fits,
-      outcome_type = Task$outcome_type,
-      conditional = Task$conditional
+      outcome_type = task$outcome_type, 
+      conditional = task$conditional
     ),
     TRUE
   )
@@ -373,8 +382,11 @@ lmtp_sdr <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
 #' @param data \[\code{data.frame}\]\cr
 #'  A \code{data.frame} in wide format containing all necessary variables
 #'  for the estimation problem. Must not be a \code{data.table}.
-#' @param trt \[\code{character}\]\cr
+#' @param trt \[\code{character}\] or \[\code{list}\]\cr
 #'  A vector containing the column names of treatment variables ordered by time.
+#'  Or, a list of vectors, the same length as the number of time points of observation.
+#'  Vectors should contain column names for the treatment variables at each time point. The list
+#'  should be ordered following the time ordering of the model.
 #' @param outcome \[\code{character}\]\cr
 #'  The column name of the outcome variable. In the case of time-to-event
 #'  analysis, a vector containing the columns names of intermediate outcome variables and the final
@@ -413,7 +425,8 @@ lmtp_sdr <- function(data, trt, outcome, baseline = NULL, time_vary = NULL,
 #'  An optional, ordered vector of the bounds for a continuous outcomes. If \code{NULL},
 #'  the bounds will be taken as the minimum and maximum of the observed data.
 #'  Should be left as \code{NULL} if the outcome type is binary.
-#' @param learners \[\code{character}\]\cr
+#' @param learners \[\code{character}\]\cr A vector of \code{mlr3superlearner} algorithms for estimation
+#'  of the outcome regression. Default is \code{c("mean", "glm")}.
 #' @param folds \[\code{integer(1)}\]\cr
 #'  The number of folds to be used for cross-fitting.
 #' @param weights \[\code{numeric(nrow(data))}\]\cr
@@ -451,11 +464,11 @@ lmtp_sub <- function(data, trt, outcome, baseline = NULL, time_vary = NULL, cens
 
   tau <- determine_tau(outcome, trt)
 
-  assertTrtCharacter(trt, tau)
+  assert_trt(trt, tau)
   checkmate::assertCharacter(cens, len = tau, null.ok = !checkmate::anyMissing(data[, outcome, drop = FALSE]))
   checkmate::assertList(time_vary, types = c("NULL", "character"), len = tau, null.ok = TRUE)
   checkmate::assertCharacter(id, len = 1, null.ok = TRUE)
-  checkmate::assertSubset(c(trt, outcome, baseline, unlist(time_vary), cens, id), names(data))
+  checkmate::assertSubset(c(unlist(trt), outcome, baseline, unlist(time_vary), cens, id), names(data))
   assertLmtpData(data, trt, outcome, baseline, time_vary, cens, id)
   assertOutcomeTypes(data, outcome, match.arg(outcome_type))
   assertReservedNames(data)
@@ -465,47 +478,49 @@ lmtp_sub <- function(data, trt, outcome, baseline = NULL, time_vary = NULL, cens
   checkmate::assertNumeric(weights, len = nrow(data), finite = TRUE, any.missing = FALSE, null.ok = TRUE)
   checkmate::assertNumber(k, lower = 0, upper = Inf)
   checkmate::assertNumber(folds, lower = 1, upper = nrow(data) - 1)
-  checkmate::assertSubset(c(trt, outcome, baseline, unlist(time_vary), cens, id), names(data))
-  checkmate::assertNumber(control$.learners_trt_folds, null.ok = TRUE)
+  checkmate::assertNumber(control$.learners_outcome_folds, null.ok = TRUE)
+  checkmate::assertNumber(control$.bound)
   checkmate::assertLogical(control$.return_full_fits, len = 1)
   checkmate::assertMatrix(conditional, null.ok = TRUE)
 
-  if(!is.null(conditional) && trt_method == "default") {
-    stop("For conditional estimation, trt_method must be set to `riesz`.")
-  }
+  task <- lmtp_task$new(
+    data = data,
+    trt = trt,
+    outcome = outcome,
+    time_vary = time_vary,
+    baseline = baseline,
+    cens = cens,
+    k = k,
+    shift = shift,
+    shifted = shifted,
+    id = id,
+    conditional = conditional,
+    outcome_type = match.arg(outcome_type),
+    V = folds,
+    weights = weights,
+    bounds = bounds,
+    bound = control$.bound
+  )
 
+  pb <- progressr::progressor(task$tau*folds)
 
-  Task <- lmtp_Task$new(data = data,
-                        trt = trt,
-                        outcome = outcome,
-                        time_vary = time_vary,
-                        baseline = baseline,
-                        cens = cens,
-                        k = k,
-                        shift = shift,
-                        shifted = shifted,
-                        id = id,
-                        conditional = conditional,
-                        outcome_type = match.arg(outcome_type),
-                        V = folds,
-                        weights = weights,
-                        bounds = bounds)
-
-  progress_bar <- progressr::progressor(Task$tau*folds)
-
-  estims <- cf_sub(Task, "tmp_lmtp_scaled_outcome", learners, control, progress_bar)
+  estims <- cf_sub(task,
+                   "tmp_lmtp_scaled_outcome",
+                   learners,
+                   control,
+                   pb)
 
   theta_sub(
     eta = list(
       m = estims$m,
-      outcome_type = Task$outcome_type,
-      bounds = Task$bounds,
-      folds = Task$folds,
-      weights = Task$weights,
+      outcome_type = task$outcome_type,
+      bounds = task$bounds,
+      folds = task$folds,
+      weights = task$weights,
       shift = if (is.null(shifted)) deparse(substitute((shift))) else NULL,
       fits_m = estims$fits,
-      outcome_type = Task$outcome_type,
-      conditional = Task$conditional
+      outcome_type = task$outcome_type,
+      conditional = task$conditional
     )
   )
 }
@@ -519,8 +534,11 @@ lmtp_sub <- function(data, trt, outcome, baseline = NULL, time_vary = NULL, cens
 #' @param data \[\code{data.frame}\]\cr
 #'  A \code{data.frame} in wide format containing all necessary variables
 #'  for the estimation problem. Must not be a \code{data.table}.
-#' @param trt \[\code{character}\]\cr
+#' @param trt \[\code{character}\] or \[\code{list}\]\cr
 #'  A vector containing the column names of treatment variables ordered by time.
+#'  Or, a list of vectors, the same length as the number of time points of observation.
+#'  Vectors should contain column names for the treatment variables at each time point. The list
+#'  should be ordered following the time ordering of the model.
 #' @param outcome \[\code{character}\]\cr
 #'  The column name of the outcome variable. In the case of time-to-event
 #'  analysis, a vector containing the columns names of intermediate outcome variables and the final
@@ -558,16 +576,17 @@ lmtp_sub <- function(data, trt, outcome, baseline = NULL, time_vary = NULL, cens
 #'  Outcome variable type (i.e., continuous, binomial, survival).
 #' @param id \[\code{character(1)}\]\cr
 #'  An optional column name containing cluster level identifiers.
-#' @param learners \[\code{character}\]\cr \bold{Only include candidate learners capable of binary classification}.
-#' @param trt_method \[\code{character}\]\cr
-#'  Method for estimating treatment assignment mechanism (default or riesz)
+#' @param learners \[\code{character}\]\cr A vector of \code{mlr3superlearner} algorithms for estimation
+#'  of the outcome regression. Default is \code{c("mean", "glm")}.
+#'  \bold{Only include candidate learners capable of binary classification}.
+#' @param riesz \[\code{logical(1)}\]\cr
+#'  Estimate treatment mechanism using riesz representer? Default is \code{FALSE}.
 #' @param folds \[\code{integer(1)}\]\cr
 #'  The number of folds to be used for cross-fitting.
 #' @param weights \[\code{numeric(nrow(data))}\]\cr
 #'  An optional vector containing sampling weights.
 #' @param control \[\code{list()}\]\cr
 #'  Output of \code{lmtp_control()}.
-#' @param ... Extra arguments. Exists for backwards compatibility.
 #'
 #' @details
 #' ## Should \code{mtp = TRUE}?
@@ -597,9 +616,8 @@ lmtp_ipw <- function(data, trt, outcome, baseline = NULL, time_vary = NULL, cens
                      k = Inf, id = NULL,
                      outcome_type = c("binomial", "continuous", "survival"),
                      learners = c("mean", "glm"),
-                     trt_method = "default",
-                     folds = 10, weights = NULL,
-                     control = lmtp_control(), ...) {
+                     riesz = FALSE, folds = 10, weights = NULL,
+                     control = lmtp_control()) {
 
   assertNotDataTable(data)
   checkmate::assertCharacter(outcome, len = if (match.arg(outcome_type) != "survival") 1,
@@ -608,11 +626,11 @@ lmtp_ipw <- function(data, trt, outcome, baseline = NULL, time_vary = NULL, cens
 
   tau <- determine_tau(outcome, trt)
 
-  assertTrtCharacter(trt, tau)
+  assert_trt(trt, tau)
   checkmate::assertCharacter(cens, len = tau, null.ok = !checkmate::anyMissing(data[, outcome, drop = FALSE]))
   checkmate::assertList(time_vary, types = c("NULL", "character"), len = tau, null.ok = TRUE)
   checkmate::assertCharacter(id, len = 1, null.ok = TRUE)
-  checkmate::assertSubset(c(trt, outcome, baseline, unlist(time_vary), cens, id), names(data))
+  checkmate::assertSubset(c(unlist(trt), outcome, baseline, unlist(time_vary), cens, id), names(data))
   assertLmtpData(data, trt, outcome, baseline, time_vary, cens, id)
   assertOutcomeTypes(data, outcome, match.arg(outcome_type))
   assertReservedNames(data)
@@ -621,65 +639,60 @@ lmtp_ipw <- function(data, trt, outcome, baseline = NULL, time_vary = NULL, cens
   checkmate::assertNumeric(weights, len = nrow(data), finite = TRUE, any.missing = FALSE, null.ok = TRUE)
   checkmate::assertNumber(k, lower = 0, upper = Inf)
   checkmate::assertNumber(folds, lower = 1, upper = nrow(data) - 1)
-  checkmate::assertSubset(c(trt, outcome, baseline, unlist(time_vary), cens, id), names(data))
-  checkmate::assertNumber(control$.learners_outcome_folds, null.ok = TRUE)
+  checkmate::assertNumber(control$.learners_trt_folds, null.ok = TRUE)
+  checkmate::assertNumber(control$.bound)
   checkmate::assertNumber(control$.trim, upper = 1)
   checkmate::assertLogical(control$.return_full_fits, len = 1)
+  check_trt_type(data, unlist(trt), mtp)
   checkmate::assertMatrix(conditional, null.ok = TRUE)
 
-  if(!is.null(conditional) && trt_method == "default") {
-    stop("For conditional estimation, trt_method must be set to `riesz`.")
+  if (!is.null(conditional) && isFALSE(riesz) {
+    stop("For conditional estimation, set `riesz = TRUE`.")
   }
 
+  task <- lmtp_task$new(
+    data = data,
+    trt = trt,
+    outcome = outcome,
+    time_vary = time_vary,
+    baseline = baseline,
+    cens = cens,
+    k = k,
+    shift = shift,
+    shifted = shifted,
+    id = id,
+    conditional = conditional,
+    outcome_type = match.arg(outcome_type),
+    V = folds,
+    weights = weights,
+    bounds = NULL,
+    bound = control$.bound
+  )
 
-  Task <- lmtp_Task$new(data = data,
-                        trt = trt,
-                        outcome = outcome,
-                        time_vary = time_vary,
-                        baseline = baseline,
-                        cens = cens,
-                        k = k,
-                        shift = shift,
-                        shifted = shifted,
-                        id = id,
-                        conditional = conditional,
-                        outcome_type = match.arg(outcome_type),
-                        V = folds,
-                        weights = weights,
-                        bounds = NULL)
-
-  progress_bar <- progressr::progressor(Task$tau*folds)
-
-  extras <- list(...)
-  if ("intervention_type" %in% names(extras)) {
-    mtp <- extras$intervention_type == "mtp"
-    warning("The `intervention_type` argument of `lmtp_ipw()` is deprecated as of lmtp 1.3.1",
-            call. = FALSE)
-  }
-
-  if(trt_method == "default") {
-    ratios <- cf_r(Task, learners, mtp, control, progress_bar)
+  pb <- progressr::progressor(task$tau*folds)
+      
+  if (isFALSE(riesz) {
+    ratios <- cf_r(task, learners, mtp, control, pb)
     ratios$ratios <- matrix(
         t(apply(ratios$ratios, 1, cumprod)),
         nrow = nrow(ratios$ratios),
         ncol = ncol(ratios$ratios)
       )
-  }
-  else {
-    ratios <- cf_rr(Task, learners, mtp, control, progress_bar)
+  } else {
+    ratios <- cf_rr(task, learners, mtp, control, pb)
   }
 
   theta_ipw(
     eta = list(
       r = ratios$ratios,
-      y = if (Task$survival) {
+      y = if (task$survival) {
         convert_to_surv(data[[final_outcome(outcome)]])
       } else {
         data[[final_outcome(outcome)]]
       },
-      folds = Task$folds,
-      weights = Task$weights,
-      tau = Task$tau,
+      folds = task$folds,
+      weights = task$weights,
+      tau = task$tau,
       shift = if (is.null(shifted)) deparse(substitute((shift))) else NULL,
       fits_r = ratios$fits,
       conditional = Task$conditional