+++ tests/testthat/test-shifted.R
@@ -15,12 +15,12 @@ sub <-
 
 ipw <-
   sw(lmtp_ipw(sim_cens, a, "Y", NULL, nodes,
-              cens, k = 0, shifted = sc, folds = 2, intervention_type = "mtp"))
+              cens, k = 0, shifted = sc, folds = 2, mtp = TRUE))
 
 tmle <-
   sw(lmtp_tmle(sim_cens, a, "Y", nodes, baseline = NULL,
                cens, k = 0, shifted = sc,
-               outcome_type = "binomial", folds = 2, intervention_type = "mtp"))
+               outcome_type = "binomial", folds = 2, mtp = TRUE))
 
 tmle_riesz <-
   sw(lmtp_tmle(sim_cens, a, "Y", nodes, baseline = NULL,
@@ -31,7 +31,7 @@ tmle_riesz <-
 sdr <-
   sw(lmtp_sdr(sim_cens, a, "Y", nodes, baseline = NULL,
               cens, k = 0, shifted = sc,
-              outcome_type = "binomial", folds = 2, intervention_type = "mtp"))
+              outcome_type = "binomial", folds = 2, mtp = TRUE))
 
 # tests
 test_that("estimator fidelity with shifted data supplied", {